Finished increase of HGT files

+++ code/primitives.h
@@ -36,7 +36,7 @@ typedef uint16_t uint16;
 typedef uint32_t uint32;
 typedef uint64_t uint64;
 
-typedef float Elevation;
+typedef int16 Elevation;
 typedef int32 Coord;
 
 // Encoded x, y offsets inside tile
